Module builtin_psl_connectors::postgres_datamodel_connector

Structs

PostgresDatamodelConnector

PostgresDatasourceProperties

Postgres-specific properties in the datasource block.

PostgresExtension

An extension defined in the extensions array of the datasource.

PostgresqlExtensions

The extensions defined in the extensions array of the datrasource.